boiler/requirements.txt
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
gunicorn==23.0.0
psycopg2-binary
Flask-Mail
Flask-Migrate
python-dotenv
flask-appbuilder
Authlib
requests